Add reproducible reorg blocks & a script for producing these blocks
- The 'import_blocks.py' should be used to first submit the blockchain_115_3677f4
which can then be followed by blockchain_116_1ccd26 (when the simple_indexer has
completed synchronizing the first batch of headers). This will reorg the chain
such that the 7 ElectrumSV transactions are all 'reorged' to chain height 116
with a new block hash.
- Fixed a bug where rows were being re-inserted erroneously (the correctness in this area is aided by the UNIQUE constraint)
- Added a blocks table for joins on confirmed_transactions table
and later will be use for the scanning API (hence the inclusion of raw_block data).